TradingMultiAgents/requirements.txt
# TradingMultiAgents Dependencies
# Core Framework
streamlit>=1.28.0,<2.0.0
streamlit-option-menu>=0.3.6
tomli>=2.0.0  # For config file validation

# LangChain and AI
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-experimental>=0.3.0,<0.4.0
langgraph>=0.0.40
openai>=1.12.0
anthropic>=0.18.0

# Financial Data APIs
finnhub-python>=2.4.18
yfinance>=0.2.18
eodhd>=1.0.0
akshare>=1.12.0
tushare>=1.4.0
stockstats>=0.6.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
typing-extensions>=4.8.0

# Async and Networking
aiohttp>=3.8.0
httpx>=0.24.0
requests>=2.31.0

# Trading Framework
backtrader>=1.9.76
pytz>=2023.3

# Database and Cache
chromadb>=0.4.0
redis>=4.5.0

# CLI and UI
typer>=0.9.0,<1.0.0
rich>=13.7.0,<14.1.0
questionary>=2.1.0,<3.0.0
chainlit>=1.0.0

# Web Scraping
praw>=7.7.1
feedparser>=6.0.10
parsel>=1.8.1

# Development
setuptools>=68.0.0
watchdog>=3.0.0
tqdm>=4.65.0
pytest>=7.4.0
black>=23.0.0

# Additional WebUI dependencies
aiofiles>=23.0.0
python-multipart>=0.0.6

# PDF generation
reportlab>=4.0.0
pypdf>=3.17.0

# Browser notifications
streamlit-push-notifications>=0.1.0
